Issue #377 refactor: 위젯 클래스 위치 변경
