+++ app/src/main/AndroidManifest.xml
@@ -16,13 +16,13 @@
         android:theme="@style/Theme.달려달력"
         android:usesCleartextTraffic="true">
         <service
-            android:name=".CalendaRunRemoteViewsService"
+            android:name=".ui.appwidget.CalendaRunRemoteViewsService"
             android:enabled="true"
             android:exported="true"
             android:permission="android.permission.BIND_REMOTEVIEWS" />
 
         <receiver
-            android:name=".CalendaRunAppWidget"
+            android:name=".ui.appwidget.CalendaRunAppWidget"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
