+++ app/src/main/java/com/drunkenboys/calendarun/ui/appwidget/CalendaRunAppWidget.kt
@@ -1,4 +1,4 @@
-package com.drunkenboys.calendarun
+package com.drunkenboys.calendarun.ui.appwidget
 
 import android.app.PendingIntent
 import android.appwidget.AppWidgetManager
@@ -8,6 +8,7 @@ import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
 import android.widget.RemoteViews
+import com.drunkenboys.calendarun.R
 import com.drunkenboys.calendarun.util.localDateToString
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
